Implement required changes
